updated with some suggestions
