+++ app/Http/Controllers/PortalController.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\EndpointRequest;
 use App\Models\Endpoint;
 use App\Models\OrganizationUser;
 use App\Services\Eherkenning\OrganizationAuthGuard;
@@ -11,8 +12,6 @@ use App\Services\HapiService;
 use Illuminate\Contracts\Auth\Guard;
 use Illuminate\Contracts\View\View;
 use Illuminate\Http\RedirectResponse;
-use Illuminate\Http\Request;
-use Illuminate\Support\Str;
 use Symfony\Component\Uid\Uuid;
 
 class PortalController extends Controller
@@ -36,9 +35,9 @@ class PortalController extends Controller
         ;
     }
 
-    public function edit(Request $request): RedirectResponse
+    public function edit(EndpointRequest $request): RedirectResponse
     {
-        $validated_data = $this->checkEndpoint($request);
+        $validated_data = $request->validated();
 
         /** @var OrganizationUser $user */
         $user = $this->guard->user();
@@ -65,39 +64,5 @@ class PortalController extends Controller
             ->route('portal.index')
             ->with('success', 'Organization information is updated successfully')
         ;
-    }
-
-    # @phpstan-ignore-next-line
-    protected function checkEndpoint(Request $request): array
-    {
-        return $request->validate([
-            'org_name' => [
-                'required',
-                'string',
-                'min:1',
-                'max:255',
-            ],
-            'id' => [
-                'string',
-                'nullable',
-                'min:0',
-                'max:255',
-                'regex:/^[a-zA-Z0-9\-]+$/',
-            ],
-            'endpoint' => [
-                'required',
-                'url',
-                'max:1024',
-                function ($attribute, $value, $fail) {
-                    $prefixes = ["https://"];
-                    if (config('app.allow_insecure_endpoints') === true) {
-                        $prefixes[] = "http://";
-                    }
-                    if (!Str::startsWith(strtolower($value), $prefixes)) {
-                        $fail($attribute . ' must start with https://');
-                    }
-                },
-            ]
-        ]);
     }
 }
